extract token validator to module
