+++ app.js
@@ -1,17 +1,16 @@
 var app        = require('express')();
 var routes     = require('./routes/director-routes');
 var bodyParser = require('body-parser');
-
-var authCtrl = require('./lib/authorization');
-var Director = require('./models/director');
-var nohm     = require('nohm').Nohm;
-var redis    = require('redis').createClient();
+var authorize  = require('./lib/authorization');
+var Director   = require('./models/director');
+var nohm       = require('nohm').Nohm;
+var redis      = require('redis').createClient();
 
 app.use(bodyParser.urlencoded({ extended: true }));
 app.use(bodyParser.json());
 
 // Introduce middleware to protect our PUT route
-app.use(authCtrl.put);
+app.use(authorize.put);
 app.use('/', routes);
 
 redis.on("connect", function() {
