+++ lib/authorization.js
@@ -1,5 +1,5 @@
-var md5 = require('MD5');
-var e   = require('../lib/error-handler');
+var e = require('../lib/error-handler');
+var tokenValidator = require('./token-validator');
 
 module.exports = {
   put: putDirectors
@@ -19,23 +19,5 @@ function putDirectors(req, res, next){
 
 
 function invalidHeader(header) {
-  return !header || badHeader(header);
-}
-
-
-function badHeader(header) {
-  var header = header.split(" ");
-  return invalidMd5(header) || invalidLength(header) || invalidFormat(header);
-}
-
-function invalidMd5(number) {
-  return !/[a-f0-9]{32}/.test(number);
-}
-
-function invalidLength(header) {
-  return header.length !== 2;
-}
-
-function invalidFormat(header) {
-  return header[0].toLowerCase() !== 'bearer';
+  return !header || tokenValidator(header);
 }
